remove timeout from TCPConnect test
This is because too many requests are currently occuring from the client

This timeout counter begins in the code, and the request is still "pending" in the browser. This will ensure the request is able to be completed.

Other work is being done to prevent the number of requests in the browser.